% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdragmap-package.R
\docType{package}
\name{Rdragmap-package}
\alias{Rdragmap}
\alias{Rdragmap-package}
\title{Rdragmap}
\description{
A source-built R wrapper for the portable, history-preserving DRAGMAP fork.
It builds and invokes one package-owned \code{dragen-os} executable through an
explicit argument vector. It does not search \code{PATH} and does not expose the
native implementation archives as an R ABI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RGenomicsETL/Rdragmap}
  \item \url{https://rgenomicsetl.github.io/Rdragmap/}
  \item Report bugs at \url{https://github.com/RGenomicsETL/Rdragmap/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Authors:
\itemize{
  \item Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}
}

Other contributors:
\itemize{
  \item Illumina, Inc. (DRAGMAP upstream implementation) [contributor, copyright holder]
  \item Mengyao Zhao (Striped Smith-Waterman implementation) [contributor]
  \item Boston College (Striped Smith-Waterman copyright holder) [copyright holder]
  \item Evan Nemerson (SIMDe copyright holder) [copyright holder]
  \item Boost Software Authors (Boost C++ Libraries authors and copyright holders) [contributor, copyright holder]
}

}
\keyword{internal}
