% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{rdragmap_is_error}
\alias{rdragmap_is_error}
\title{Test whether a value is an Rdragmap operational error}
\usage{
rdragmap_is_error(value)
}
\arguments{
\item{value}{Any R value.}
}
\value{
A logical scalar.
}
\description{
Test whether a value is an Rdragmap operational error
}
