% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executable.R
\name{RdragmapExecutables}
\alias{RdragmapExecutables}
\title{Explicit dragen-os child executable}
\usage{
RdragmapExecutables(dragen_os = character(0))
}
\arguments{
\item{dragen_os}{Absolute path to \code{dragen-os}.}
}
\description{
Explicit dragen-os child executable
}
